trace.c
